+++ src/pxs_core/pxs_pxs.rs
@@ -1,0 +1,7 @@
+use crate::{pxs_addmod, pxs_newmod};
+
+/// Initialize `pxs_pxs` module.
+pub(crate) fn init() {
+    let pxs_pxs = pxs_newmod(c"pxs_pxs".as_ptr());
+    pxs_addmod(pxs_pxs);
+}