+++ src/yoyo/mod.rs
@@ -1,3 +1,0 @@
-pub(crate) mod yoyo {
-    include!(concat!(env!("OUT_DIR"), "/yoyo_bindings.rs"));
-}