+++ CHANGELOG.md
@@ -68,7 +68,10 @@
 - Added `test_bytes.rs`
 - Added `test_yoyo.rs`
 - removed `php` support. (Was never completed and it just got in the way.) I will add it back as a optional C++ module in 0.8
-- Added `tst_exception_in_chunk.rs`
+- Added `test_exception_in_chunk.rs`
 - Updated doc comment for `pxs_newcopy` to be explicit about how the deleter is moved.
 - Fixed module naming for grandchildren.
 - Removed `c_tests` directory
+- Added `cargo-zigbuild` support.
+- Added `test_grandparent.rs` 
+- Added `pxs_Exception` to `pxs_varsize`. Uses the same logic as `pxs_String`